major updates following Alex's changes

+++ bin/collect_metadata.sh
@@ -12,13 +12,22 @@ then
   ## download the so-called soft_family file, and use it to generate same files as above
   PAD=`echo $SERIES | perl -ne 's/\d{3}$/nnn/; print'`
   wget -O ${SERIES}_family.soft.gz https://ftp.ncbi.nlm.nih.gov/geo/series/$PAD/$SERIES/soft/${SERIES}_family.soft.gz
+  ## -f overwrites the old stuff
   gzip -fd ${SERIES}_family.soft.gz
+  if [[ ! -s ${SERIES}_family.soft ]]
+  then
+    >&2 echo "ERROR: Failed to download ${SERIES}_family.soft file; please make sure the series you requested exists, or fix the download URL!"
+    exit 1
+  fi
 
   ## samples here are GSM IDs; usually for a 10x GSM==SRS==SRX, but I haven't checked *all* of the SRA you know 
   grep Sample_geo_accession ${SERIES}_family.soft | awk '{print $3}' | sort | uniq > $SERIES.sample.list
   grep Series_relation ${SERIES}_family.soft | perl -ne 'print "$1\n" if (m/(PRJ[A-Z]+\d+)/)' | sort | uniq > $SERIES.project.list
-  
+ 
+  ## first variable is used to spot dbGap and other problematic datasets;
+  ## second variable is used to find SubSeries when SuperSeries does not produce any meaningful ENA links
   EXPIDS=`grep Series_relation ${SERIES}_family.soft | grep -v PRJ | wc -l`
+  SUBGSE=`grep Series_relation ${SERIES}_family.soft | grep SuperSeries | perl -ne 'print "$1\n" if (m/(GSE\d+)/)'` 
   
   ## few sanity checks:
   if [[ `cat $SERIES.project.list | wc -l` != "1" ]]
@@ -33,10 +42,49 @@ then
 
   ## curl info about each run (SRR/ERR/DRR) from ENA API; v2 pulls GSM data etc 
   RET=1
-  until [ ${RET} -eq 0 ]
+  TRIES=1
+  until (( $RET == 0 ))
   do
     ./curl_ena_metadata.sh $SERIES.project.list > $SERIES.ena.tsv 
     RET=$?
+
+    ## this either pulls sub-series data (and replaces $SERIES.project.list with useful PRJNA* IDs), or just quits after 5 tries
+    TRIES=$((TRIES+1))
+    if (( $TRIES > 5 )) 
+    then
+      >&2 echo "WARNING: No ENA records can be retrieved for GEO projects listed in $SERIES.project.list!"
+      if [[ $SUBGSE == "" ]]
+      then 
+        >&2 echo "ERROR: No GSE subseries were listed in ${SERIES}_family.soft - no alternative PRJNA* to be found, and no ENA entries can be retrieved!"
+        exit 1
+      else 
+        >&2 echo "WARNING: replacing $SERIES.project.list with sub-series projects.."
+        rm $SERIES.project.list 
+        for i in $SUBGSE
+        do
+          PAD=`echo $i | perl -ne 's/\d{3}$/nnn/; print'`
+          wget -O ${i}_family.soft.gz https://ftp.ncbi.nlm.nih.gov/geo/series/$PAD/$i/soft/${i}_family.soft.gz
+          gzip -fd ${i}_family.soft.gz
+          grep Series_relation ${i}_family.soft | perl -ne 'print "$1\n" if (m/(PRJ[A-Z]+\d+)/)' | sort | uniq >> $SERIES.project.list
+        done
+           
+        ## now, once we re-populated $SERIES.project.list, let's try to get ENA records for the new IDs.. 
+        >&2 echo "WARNING: pulling ENA records using sub-series project identifiers.."
+        RET=1
+        TRIES=1
+        until (( $RET == 0 ))
+        do
+          ./curl_ena_metadata.sh $SERIES.project.list > $SERIES.ena.tsv
+          RET=$?
+          TRIES=$((TRIES+1))
+          if (( $TRIES > 5 ))
+          then
+            >&2 echo "ERROR: Still no ENA records can be retrieved for the GEO SUBSERIES projects listed in $SERIES.project.list, I quit!"
+            exit 1
+          fi
+        done
+      fi
+    fi 
     sleep 1
   done
 
@@ -62,16 +110,30 @@ elif [[ $SERIES == E-MTAB* ]]
 then
   ## sdrf's are wonderful, but don't have the project ID, which is *annoying*. That's OK though, we'll go by SRS. 
   wget -O $SERIES.sdrf.txt https://www.ebi.ac.uk/biostudies/files/$SERIES/$SERIES.sdrf.txt
+  wget -O $SERIES.idf.txt https://www.ebi.ac.uk/biostudies/files/$SERIES/$SERIES.idf.txt
+  if [[ ! -s $SERIES.sdrf.txt ]] 
+  then
+    >&2 echo "ERROR: Failed to download $SERIES.sdrf.txt file; please make sure the series you requested exists, or fix the download URL!"
+    exit 1
+  fi 
 
-  ## samples are ERS in case of ArrayExpress 
+  ## samples are ERS in case of ArrayExpress. Why not ERX, you might ask? Yes, ask you might.   
   cat $SERIES.sdrf.txt | tr '\t' '\n' | grep "^ERS" | sort | uniq > $SERIES.sample.list
 
   ## curl info about each run (SRR/ERR/DRR) from ENA API; in this case we use ERS IDs 
   RET=1
-  until [ ${RET} -eq 0 ]
+  TRIES=1
+  until (( $RET == 0 ))
   do
+    ## for ArrayExpress, we query by sample ID because sdrf doesn't list the BioProject ID
     ./curl_ena_metadata.sh $SERIES.sample.list > $SERIES.ena.tsv 
     RET=$?
+    TRIES=$((TRIES+1))
+    if (( $TRIES > 5 )) 
+    then
+      >&2 echo "ERROR: No ENA records can be retrieved for ArrayExpress samples $SERIES.sample.list!"
+      exit 1
+    fi 
     sleep 1
   done
 
@@ -99,10 +161,17 @@ then
 
   ## curl info about each run (SRR/ERR/DRR) from ENA API; v2 pulls GSM data etc 
   RET=1
-  until [ ${RET} -eq 0 ]
+  TRIES=1
+  until (( $RET == 0 ))
   do
     ./curl_ena_metadata.sh $SERIES.project.list > $SERIES.ena.tsv 
     RET=$?
+    TRIES=$((TRIES+1))
+    if (( $TRIES > 5 )) 
+    then
+      >&2 echo "ERROR: No ENA records can be retrieved for BioProject(s) $SERIES.project.list!"
+      exit 1
+    fi 
     sleep 1
   done
